Refactor strategy into a separate file
